Centralize logging and consolidate helpers
