+++ custom_exceptions.py
@@ -1,25 +1,19 @@
 # custom_exceptions.py - Custom exception classes for KYO QA Tool
 
-class KYOQAToolError(Exception):
-    """Base exception for all KYO QA Tool errors."""
-    pass
+from utils import (
+    KYOQAToolError,
+    FileLockError,
+    ExcelGenerationError,
+    PDFExtractionError,
+    PatternMatchError,
+    ConfigurationError,
+)
 
-class FileLockError(KYOQAToolError):
-    """Raised when a file is locked by another process."""
-    pass
-
-class ExcelGenerationError(KYOQAToolError):
-    """Raised when Excel file generation fails."""
-    pass
-
-class PDFExtractionError(KYOQAToolError):
-    """Raised when PDF text extraction fails."""
-    pass
-
-class PatternMatchError(KYOQAToolError):
-    """Raised when pattern matching fails."""
-    pass
-
-class ConfigurationError(KYOQAToolError):
-    """Raised when there's a configuration issue."""
-    pass
+__all__ = [
+    "KYOQAToolError",
+    "FileLockError",
+    "ExcelGenerationError",
+    "PDFExtractionError",
+    "PatternMatchError",
+    "ConfigurationError",
+]
