+++ custom_patterns.py
@@ -1,32 +1,15 @@
 # custom_patterns.py
 # This file stores user-defined regex patterns.
 
-MODEL_PATTERNS = [
-    r'\b(TASKalfa|ECOSYS)\s*[\w-]+\b',
-    r'\b(PF|DF|MK|AK|DP|BF|JS)-\d+[\w-]*\b',
-    r'\bFS-C?\d+[a-zA-Z]*\b',
-    r'\bKM-\d+[a-zA-Z]*\b',
-    r'\bCS\s\d+[a-zA-Z]*\b',
-    r'\bTASKalfa\s+\d+(ci|cxi|cx)\b',
-    r'\bECOSYS\s+M\d+(cdn|cdnl|ci|cidn)\b',
-    r'\bECOSYS\s+P\d+(dn|dtn|d)\b',
-    r'\bDP-\d+\b',
-    r'\bM\d+idn\b',
-    r'\bM\d+idnf\b',
-    r'\bVi\d+\b',
-    r'\b[A-Z]{2,}-\d{3,}\b',
-    r'\b[A-Z]{3,}\s[A-Z]\d{4}[a-z]*\b',
-    r'\bP\d+cdn\b',
-    r'\bP\d+d\b',
-    r'\bP\d+dn\b',
-    r'\bM\d+dn\b',
+from utils import (
+    MODEL_PATTERNS,
+    QA_NUMBER_PATTERNS,
+    PART_NUMBER_PATTERNS,
+)
+
+__all__ = [
+    "MODEL_PATTERNS",
+    "QA_NUMBER_PATTERNS",
+    "PART_NUMBER_PATTERNS",
 ]
 
-QA_NUMBER_PATTERNS = [
-    r'\bQA[-_]?\d+[\w-]*\b',
-    r'\bSB[-_]?\d+[\w-]*\b',
-]
-
-PART_NUMBER_PATTERNS = [
-    r'\b\d{2,}[A-Z]{1,}\d{5,}\b',
-]
